+++ src/Frontend/Dfe.Complete/Models/BaseProjectsPageModel.cs
@@ -22,4 +22,7 @@ public abstract class BaseProjectsPageModel(string currentNavigation) : PageMode
 
     public static string GetProjectSummaryUrl(ProjectId projectId) =>
         string.Format(RouteConstants.ProjectTaskList, projectId.Value);
+
+    protected IActionResult HasPageFound(bool condition) =>
+        condition ? StatusCode(404) : null!;
 }